usersTable created as a component.
